+++ dap-etl/transformation/src/main/scala/org/broadinstitute/monster/dap/CslbTransformations.scala
@@ -1,5 +1,6 @@
 package org.broadinstitute.monster.dap
 
+import org.broadinstitute.monster.dap.common.RawRecord
 import org.broadinstitute.monster.dogaging.jadeschema.table.Cslb
 
 object CslbTransformations {
