+++ dap-etl/transformation/src/main/scala/org/broadinstitute/monster/dap/EnvironmentTransformationPipelineBuilder.scala
@@ -4,6 +4,7 @@ import com.spotify.scio.ScioContext
 import com.spotify.scio.values.SCollection
 import org.broadinstitute.monster.common.{PipelineBuilder, StorageIO}
 import org.broadinstitute.monster.common.msg._
+import org.broadinstitute.monster.dap.common.{Args, RawRecord}
 import org.slf4j.{Logger, LoggerFactory}
 import upack.Msg
