+++ dap-etl/transformation/src/main/scala/org/broadinstitute/monster/dap/EnvironmentTransformations.scala
@@ -2,6 +2,7 @@ package org.broadinstitute.monster.dap
 
 import org.broadinstitute.monster.dap.environment._
 import org.broadinstitute.monster.dap.HLESurveyTransformationPipelineBuilder.logger
+import org.broadinstitute.monster.dap.common.RawRecord
 import org.broadinstitute.monster.dogaging.jadeschema.table.Environment
 
 object EnvironmentTransformations {
