+++ dap-etl/transformation/src/main/scala/org/broadinstitute/monster/dap/HLESurveyTransformationPipeline.scala
@@ -2,6 +2,7 @@ package org.broadinstitute.monster.dap
 
 import com.spotify.scio.ScioResult
 import org.broadinstitute.monster.common.{PipelineBuilder, ScioApp}
+import org.broadinstitute.monster.dap.common.{Args, PostProcess}
 
 /** Entry-point for the HLE transformation pipeline. */
 object HLESurveyTransformationPipeline extends ScioApp[Args] {
