+++ dap-etl/transformation/src/main/scala/org/broadinstitute/monster/dap/environment/WalkabilityTransformations.scala
@@ -1,8 +1,8 @@
 package org.broadinstitute.monster.dap.environment
 
-import java.time.LocalDate
+import org.broadinstitute.monster.dap.common.RawRecord
 
-import org.broadinstitute.monster.dap.RawRecord
+import java.time.LocalDate
 import org.broadinstitute.monster.dogaging.jadeschema.fragment.EnvironmentWalkability
 
 object WalkabilityTransformations {
