+++ dap-etl/transformation/src/test/scala/org/broadinstitute/monster/dap/EnvironmentTransformationsSpec.scala
@@ -1,5 +1,6 @@
 package org.broadinstitute.monster.dap
 
+import org.broadinstitute.monster.dap.common.RawRecord
 import org.broadinstitute.monster.dogaging.jadeschema.fragment._
 import org.broadinstitute.monster.dogaging.jadeschema.table.Environment
 import org.scalatest.flatspec.AnyFlatSpec
