+++ dap-etl/transformation/src/test/scala/org/broadinstitute/monster/dap/environment/CensusTransformationsSpec.scala
@@ -1,6 +1,6 @@
 package org.broadinstitute.monster.dap.dog
 
-import org.broadinstitute.monster.dap.RawRecord
+import org.broadinstitute.monster.dap.common.RawRecord
 import org.broadinstitute.monster.dap.environment.CensusTransformations
 import org.scalatest.OptionValues
 import org.scalatest.flatspec.AnyFlatSpec
